streamlit/requirements.txt
streamlit
mysql-connector-python
pandas
requests
python-dotenv
